walletService/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astor==0.8.1
astroid==2.4.2
build==0.4.0
cached-property==1.5.2
certifi==2020.6.20
color==0.1
distlib==0.3.1
Django==3.2.11
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==21.1
django-six==1.0.4
djangorestframework==3.13.1
docutils==0.17.1
enum34==1.1.10
filelock==3.0.12
flake8==4.0.1
ghostscript==0.7
importlib-metadata==4.2.0
importlib-resources==5.1.4
isort==5.7.0
jeepney==0.6.0
lazy-object-proxy==1.4.3
Markdown==3.3.4
mccabe==0.6.1
numpy==1.19.5
password-endorser-sarath-rajan==0.0.1
pep517==0.10.0
pkginfo==1.7.0
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.6.0
pyparsing==2.4.7
python3-ghostscript==0.5.0
pytz==2021.3
readme-renderer==29.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
screen==1.0.1
SecretStorage==3.3.1
sqlparse==0.4.2
term==2.3
termcolor==1.1.0
toml==0.10.2
twine==3.4.1
typed-ast==1.4.2
typing_extensions==4.0.1
virtualenv==20.4.7
webencodings==0.5.1
xlwt==1.3.0
zipp==3.6.0
